+++ x/accounts/internal/implementation/context.go
@@ -6,6 +6,7 @@ import (
 
 	"cosmossdk.io/collections"
 	"cosmossdk.io/core/address"
+	"cosmossdk.io/core/header"
 	"cosmossdk.io/core/store"
 	"cosmossdk.io/x/accounts/internal/prefixstore"
 )
@@ -53,8 +54,8 @@ func MakeAccountContext(
 		sender:            sender,
 		whoami:            accountAddr,
 		originalContext:   ctx,
-		moduleExecUntyped: moduleExecUntyped,
 		moduleExec:        moduleExec,
+		moduleExecUntyped: moduleExecUntyped,
 		moduleQuery:       moduleQuery,
 	})
 }
@@ -109,8 +110,8 @@ func QueryModule[Resp any, RespProto ProtoMsgG[Resp], Req any, ReqProto ProtoMsg
 	return resp, nil
 }
 
-// OpenKVStore returns the prefixed store for the account given the context.
-func OpenKVStore(ctx context.Context) store.KVStore {
+// openKVStore returns the prefixed store for the account given the context.
+func openKVStore(ctx context.Context) store.KVStore {
 	return ctx.Value(contextKey{}).(contextValue).store
 }
 
@@ -124,7 +125,9 @@ func Whoami(ctx context.Context) []byte {
 	return ctx.Value(contextKey{}).(contextValue).whoami
 }
 
-// OriginalContext returns the original context for the account given the context.
-func OriginalContext(ctx context.Context) context.Context {
-	return ctx.Value(contextKey{}).(contextValue).originalContext
+type headerService struct{ header.Service }
+
+func (h headerService) GetHeaderInfo(ctx context.Context) header.Info {
+	originalContext := ctx.Value(contextKey{}).(contextValue).originalContext
+	return h.Service.GetHeaderInfo(originalContext)
 }