+++ x/accounts/internal/implementation/implementation.go
@@ -6,6 +6,7 @@ import (
 
 	"cosmossdk.io/collections"
 	"cosmossdk.io/core/address"
+	"cosmossdk.io/core/header"
 	"github.com/cosmos/cosmos-sdk/codec"
 )
 
@@ -14,6 +15,7 @@ type Dependencies struct {
 	SchemaBuilder *collections.SchemaBuilder
 	AddressCodec  address.Codec
 	BinaryCodec   codec.Codec
+	HeaderService header.Service
 }
 
 // AccountCreatorFunc is a function that creates an account.
@@ -21,14 +23,20 @@ type AccountCreatorFunc = func(deps Dependencies) (string, Account, error)
 
 // MakeAccountsMap creates a map of account names to account implementations
 // from a list of account creator functions.
-func MakeAccountsMap(addressCodec address.Codec, binaryCodec codec.Codec, accounts []AccountCreatorFunc) (map[string]Implementation, error) {
+func MakeAccountsMap(
+	addressCodec address.Codec,
+	binaryCodec codec.Codec,
+	hs header.Service,
+	accounts []AccountCreatorFunc,
+) (map[string]Implementation, error) {
 	accountsMap := make(map[string]Implementation, len(accounts))
 	for _, makeAccount := range accounts {
-		stateSchemaBuilder := collections.NewSchemaBuilderFromAccessor(OpenKVStore)
+		stateSchemaBuilder := collections.NewSchemaBuilderFromAccessor(openKVStore)
 		deps := Dependencies{
 			SchemaBuilder: stateSchemaBuilder,
 			AddressCodec:  addressCodec,
 			BinaryCodec:   binaryCodec,
+			HeaderService: headerService{hs},
 		}
 		name, accountInterface, err := makeAccount(deps)
 		if err != nil {