+++ x/accounts/keeper.go
@@ -16,6 +16,7 @@ import (
 	"cosmossdk.io/core/address"
 	"cosmossdk.io/core/branch"
 	"cosmossdk.io/core/event"
+	"cosmossdk.io/core/header"
 	"cosmossdk.io/core/store"
 	"cosmossdk.io/x/accounts/accountstd"
 	"cosmossdk.io/x/accounts/internal/implementation"
@@ -55,9 +56,6 @@ type SignerProvider interface {
 	GetMsgV1Signers(msg gogoproto.Message) ([][]byte, proto.Message, error)
 }
 
-// BranchExecutor defines an interface used to execute ops in a branch.
-type BranchExecutor = branch.Service
-
 type InterfaceRegistry interface {
 	RegisterInterface(name string, iface any, impls ...gogoproto.Message)
 	RegisterImplementations(iface any, impls ...gogoproto.Message)
@@ -66,7 +64,8 @@ type InterfaceRegistry interface {
 func NewKeeper(
 	ss store.KVStoreService,
 	es event.Service,
-	bs BranchExecutor,
+	hs header.Service,
+	bs branch.Service,
 	addressCodec address.Codec,
 	signerProvider SignerProvider,
 	execRouter MsgRouter,
@@ -79,13 +78,12 @@ func NewKeeper(
 	keeper := Keeper{
 		storeService:    ss,
 		eventService:    es,
-		branchExecutor:  bs,
 		addressCodec:    addressCodec,
-		signerProvider:  signerProvider,
+		branchExecutor:  bs,
 		msgRouter:       execRouter,
+		signerProvider:  signerProvider,
 		queryRouter:     queryRouter,
 		Schema:          collections.Schema{},
-		codec:           codec,
 		AccountNumber:   collections.NewSequence(sb, AccountNumberKey, "account_number"),
 		AccountsByType:  collections.NewMap(sb, AccountTypeKeyPrefix, "accounts_by_type", collections.BytesKey, collections.StringValue),
 		AccountByNumber: collections.NewMap(sb, AccountByNumber, "account_by_number", collections.BytesKey, collections.Uint64Value),
@@ -97,7 +95,7 @@ func NewKeeper(
 		return Keeper{}, err
 	}
 	keeper.Schema = schema
-	keeper.accounts, err = implementation.MakeAccountsMap(keeper.addressCodec, keeper.codec, accounts)
+	keeper.accounts, err = implementation.MakeAccountsMap(keeper.addressCodec, hs, keeper.codec, accounts)
 	if err != nil {
 		return Keeper{}, err
 	}
@@ -110,7 +108,7 @@ type Keeper struct {
 	storeService   store.KVStoreService
 	eventService   event.Service
 	addressCodec   address.Codec
-	branchExecutor BranchExecutor
+	branchExecutor branch.Service
 	msgRouter      MsgRouter
 	signerProvider SignerProvider
 	queryRouter    QueryRouter